backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
django-environ==0.11.2
Pillow==11.3.0
geopy==2.4.0
shapely==2.1.1 